# Improve ToString/ToStringE performance