--- a/caste.go
+++ b/caste.go
@@ -919,30 +919,8 @@ func indirect(a interface{}) interface{} {
 	return v.Interface()
 }
 
-// From html/template/content.go
-// Copyright 2011 The Go Authors. All rights reserved.
-// indirectToStringerOrError returns the value, after dereferencing as many times
-// as necessary to reach the base type (or nil) or an implementation of fmt.Stringer
-// or error,
-func indirectToStringerOrError(a interface{}) interface{} {
-	if a == nil {
-		return nil
-	}
-
-	errorType := reflect.TypeOf((*error)(nil)).Elem()
-	fmtStringerType := reflect.TypeOf((*fmt.Stringer)(nil)).Elem()
-
-	v := reflect.ValueOf(a)
-	for !v.Type().Implements(fmtStringerType) && !v.Type().Implements(errorType) && v.Kind() == reflect.Ptr && !v.IsNil() {
-		v = v.Elem()
-	}
-	return v.Interface()
-}
-
 // ToStringE casts an interface to a string type.
 func ToStringE(i interface{}) (string, error) {
-	i = indirectToStringerOrError(i)
-
 	switch s := i.(type) {
 	case string:
 		return s, nil
@@ -993,6 +971,11 @@ func ToStringE(i interface{}) (string, error) {
 	case error:
 		return s.Error(), nil
 	default:
+		v := reflect.ValueOf(s)
+		for v.Kind() == reflect.Ptr && !v.IsNil() {
+			return ToStringE(v.Elem().Interface())
+		}
+
 		return "", fmt.Errorf("unable to cast %#v of type %T to string", i, i)
 	}
 }