Currently `ToString`/`ToStringE` always attempts to perform pointer dereference before the type checks which adds significant overhead when casting non-pointer values (_for pointers it is also not very optimal because the reflection of `errorType` and `fmtStringerType` would be evaluated on every `ToString`/`ToStringE` call_).


This PR:
- moves the pointer check in the `default` clause and calls `ToStringE` with its underlying value
- updates the existing tests with equivalent checks for pointers

Below are some basic benchmark results from my local environment (_the best from 3 consequent runs_):

1. **Plain value**
    ```go
    func BenchmarkToString(b *testing.B) {
        for range b.N {
            cast.ToString("test")
        }
    }
    ```
    ```
    // Old:
    BenchmarkToString-4     19877439          60.69 ns/op
    
    // New:
    BenchmarkToString-4     282848110         4.364 ns/op
    ```

2. **Pointer**
    ```go
    func BenchmarkToString(b *testing.B) {
        v := "test"
        for range b.N {
            cast.ToString(&v)
        }
    }
    ```
    ```
    // Old:
    BenchmarkToString-4      4797831         287.2 ns/op

    // New:
    BenchmarkToString-4     10021204         216.1 ns/op
    ```